Renamed crud by store describes better that these are operations on the stored data i.e. db
